refactoring: rearrange types
